Unify the eeprom functions across Motherboard and Extruder controller
