+++ v2/src/Extruder/Main.cc
@@ -23,6 +23,7 @@
 #include "ExtruderMotor.hh"
 #include "ThermistorTable.hh"
 #include <avr/interrupt.h>
+#include "Eeprom.hh"
 #include "EepromMap.hh"
 #include "ExtruderBoard.hh"
 #include "MotorController.hh"
