+++ v2/src/Extruder/boards/ecv34/CoolingFan.cc
@@ -1,5 +1,6 @@
 #include "CoolingFan.hh"
 #include "ExtruderMotor.hh"
+#include "Eeprom.hh"
 #include "EepromMap.hh"
 
 #define FAN_ENABLED 1
